공지사항 매뉴 신설

+++ src/main/resources/static/js/app/index.js
@@ -24,6 +24,9 @@ const main = {
     },
 
     save() {
+
+        const writeTo = $("body").data("writeTo")
+
         const data = {
             title: $("#input-title").val(),
             author: $("#input-author").val(),
@@ -33,13 +36,13 @@ const main = {
 
         $.ajax({
             type: "POST",
-            url: "/api/v1/posts",
+            url: "/api/v1/" + writeTo,
             dataType: "json",
             contentType: "application/json; charset=utf-8",
             data: JSON.stringify(data)
         }).done(() => {
             alert("글이 등록되었습니다.")
-            window.location.href = "/"
+            window.location.href = "/" + (writeTo != "posts" ? writeTo : "")
         }).fail(err => {
             alert(JSON.stringify(err))
         })
@@ -47,6 +50,9 @@ const main = {
     },
 
     update() {
+
+        const writeTo = $("body").data("writeTo")
+
         const data = {
             title: $("#input-title").val(),
             content: $("#txt-content").val()
@@ -56,13 +62,13 @@ const main = {
 
         $.ajax({
             type: "PUT",
-            url: "/api/v1/posts/" + id,
+            url: "/api/v1/" + writeTo + "/" + id,
             dataType: "json",
             contentType: "application/json; charset=utf-8",
             data: JSON.stringify(data)
         }).done(() => {
             alert("글이 수정되었습니다.")
-            window.location.href = "/"
+            window.location.href = "/" + (writeTo != "posts" ? writeTo : "")
         }).fail(err => {
             alert(JSON.stringify(err))
             console.log(err)
@@ -72,16 +78,21 @@ const main = {
     delete() {
         
         const id = $("#input-id").val()
+        if(!confirm("정말 이 글을 삭제하시겠습니까?")) {
+            return false
+        }
+
+        const writeTo = $("body").data("writeTo")
         
         $.ajax({
             type: "DELETE",
-            url: "/api/v1/posts/" + id,
+            url: "/api/v1/" + writeTo + "/" + id,
             dataType: "json",
             contentType: "application/json; charset=utf-8"
         }).done(res => {
             console.log(res)
             alert(`${res}번 글이 삭제되었습니다.`)
-            window.location.href = "/"
+            window.location.href = "/" + (writeTo != "posts" ? writeTo : "")
         }).fail(err => {
             alert(JSON.stringify(err))
             console.log(err)
